+++ Plandomizer.py
@@ -17,7 +17,7 @@ from Fill import FillError
 from Hints import gossipLocations, GossipText, hint_func
 from rs.hints import HintArea
 from Item import ItemFactory, ItemInfo, ItemIterator, is_item, Item
-from ItemPool import item_groups, get_junk_item, song_list, trade_items, child_trade_items, eggs, triforce_blitz_items, triforce_pieces
+from ItemPool import item_groups, get_junk_item, song_list, trade_items, child_trade_items, ocarina_buttons, eggs, triforce_blitz_items, triforce_pieces
 from JSONDump import dump_obj, CollapseList, CollapseDict, AlignedDict, SortedDict
 from Location import Location, LocationIterator, LocationFactory
 from LocationList import location_groups, location_table
@@ -47,6 +47,7 @@ per_world_keys = (
     'songs',
     'entrances',
     'locations',
+    ':randomized_starting_items',
     ':skipped_locations',
     ':woth_locations',
     ':goal_locations',
@@ -274,6 +275,7 @@ class WorldDistribution:
         self.item_pool: Optional[dict[str, ItemPoolRecord]] = None
         self.entrances: Optional[dict[str, EntranceRecord]] = None
         self.locations: Optional[dict[str, LocationRecord | list[LocationRecord]]] = None
+        self.randomized_starting_items: Optional[dict[str, int]] = None
         self.woth_locations: Optional[dict[str, LocationRecord]] = None
         self.goal_locations: Optional[dict[str, dict[str, dict[str, LocationRecord | dict[str, LocationRecord]]]]] = None
         self.barren_regions: Optional[list[str]] = None
@@ -302,6 +304,7 @@ class WorldDistribution:
             'item_pool': {name: ItemPoolRecord(record) for (name, record) in src_dict.get('item_pool', {}).items()},
             'entrances': {name: EntranceRecord(record) for (name, record) in src_dict.get('entrances', {}).items()},
             'locations': {name: [LocationRecord(rec) for rec in record] if is_pattern(name) else LocationRecord(record) for (name, record) in src_dict.get('locations', {}).items() if not is_output_only(name)},
+            'randomized_starting_items': None,
             'woth_locations': None,
             'goal_locations': None,
             'barren_regions': None,
@@ -334,6 +337,7 @@ class WorldDistribution:
             'item_pool': SortedDict({name: record.to_json() for (name, record) in self.item_pool.items()}),
             'entrances': {name: record.to_json() for (name, record) in self.entrances.items()},
             'locations': {name: [rec.to_json() for rec in record] if is_pattern(name) else record.to_json() for (name, record) in self.locations.items()},
+            ':randomized_starting_items': None if self.randomized_starting_items is None else {name: count for (name, count) in self.randomized_starting_items.items()},
             ':skipped_locations': {loc.name: LocationRecord.from_item(loc.item).to_json() for loc in self.skipped_locations},
             ':woth_locations': None if self.woth_locations is None else {name: record.to_json() for (name, record) in self.woth_locations.items()},
             ':goal_locations': self.goal_locations,
@@ -1078,6 +1082,10 @@ class WorldDistribution:
             if name in triforce_pieces or record.count == 0:
                 continue
             save_context.give_item(world, name, record.count)
+
+    def give_randomized_items(self, world: World, save_context: SaveContext) -> None:
+        for item, count in world.randomized_starting_items.items():
+            save_context.give_item(world, item, count)
 
     def get_starting_item(self, item: str) -> int:
         items = self.settings.starting_items